% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatches.R
\name{countMatches}
\alias{countMatches}
\title{Converts position output to logical match output.}
\usage{
countMatches(matches, peaks, out = c("counts", "matches"))
}
\arguments{
\item{matches}{Output of \code{\link{motifmatchr::matchMotifs}(..., out='positions')}}

\item{peaks}{Subject input for \code{matchMotifs}.}

\item{out}{One of \code{'counts'} or \code{matches}.}
}
\value{
A counts matrix with motifs as columns and peaks as rows.
}
\description{
Converts position output to logical match output.
}
