% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatches.R
\name{getMatches}
\alias{getMatches}
\title{Searches a given set of genomic ranges for the motifs returned by mergeMotifs.
The background frequency is obtained from the total nucleotide frequencies in peaks.}
\usage{
getMatches(peaks, genome, out = "matches", motifs = CrobustaMotifs, ...)
}
\arguments{
\item{peaks}{A GenomicRanges or other subject input for matchMotifs.}

\item{genome}{A BSgenome object or other genome input for matchMotifs.}

\item{out}{Type of output to return.}

\item{motifs}{A PWMatrixList.}

\item{...}{Additional arguments to \code{matchMotifs()}}
}
\value{
A \code{matches} object.
}
\description{
Searches a given set of genomic ranges for the motifs returned by mergeMotifs.
The background frequency is obtained from the total nucleotide frequencies in peaks.
}
\seealso{
\code{\link{mergeMotifs}}, \code{\link{motifmatchr::matchMotifs}}
}
