% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKmers.R
\name{getMatrices}
\alias{getMatrices}
\title{Returns position weight matrices of all permutations of some NT length k.}
\usage{
getMatrices(k, bases = c("A", "C", "G", "T"))
}
\arguments{
\item{k}{bp length of each mer.}
}
\value{
A list of logical matrices.
}
\description{
Returns position weight matrices of all permutations of some NT length k.
}
\seealso{
\code{\link{PWMatrixList}}
}
