% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatches.R
\name{getP}
\alias{getP}
\title{Convert motifmatchr score to p-value.}
\usage{
getP(mat, scores, bg)
}
\arguments{
\item{mat}{A position weight matrix}

\item{scores}{A vector of motif scores}

\item{bg}{Nucleotide background frequencies}
}
\value{
A p-value for the motif match
}
\description{
Convert motifmatchr score to p-value.
}
\seealso{
\code{\link{TFMPvalue::TFMs2pv}}
}
