% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatches.R
\name{matchPois}
\alias{matchPois}
\title{Accepts a vector of peak IDs, and performs a Poisson test for enrichmet of motifs in \code{peaks} versus a background of all elements in \code{matches}.}
\usage{
matchPois(peaks, matches, padj.method = "fdr", allow.duplicates = F)
}
\arguments{
\item{peaks}{A character vector of peak IDs corresponding to the \code{row.names} of \code{matches}.}

\item{matches}{An output from motifmatchr::matchMotifs.}

\item{padj.method}{The method to be used by \code{p.adjust()}.}

\item{allow.duplicates}{Whether to remove duplicates from \code{peaks}.}
}
\value{
A \code{data.frame} containing the test results.
}
\description{
Accepts a vector of peak IDs, and performs a Poisson test for enrichmet of motifs in \code{peaks} versus a background of all elements in \code{matches}.
}
\seealso{
\code{\link{getMatches}}, \code{\link{motifmatchr::matchMotifs}}
}
