% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatches.R
\name{regMat}
\alias{regMat}
\title{Determines a matrix of potential regulatory TFs.}
\usage{
regMat(geneToPeak, matches)
}
\arguments{
\item{geneToPeak}{A 2-column matrix with the first row corresponding to the \code{row.names} of \code{matches} and the second giving the Gene ID.}

\item{matches}{An output from \code{matchMotifs}.}
}
\value{
A logical matrix with columns corresponding to the motifs in \code{matches} and rows corresponding to the genes in \code{geneToPeak}.
}
\description{
Determines a matrix of potential regulatory TFs.
}
\seealso{
\code{\link{getOverlaps}}, \code{\link{matchMotifs}}
}
