% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMatches.R
\name{setP}
\alias{setP}
\title{Add p-value to GRangesList with a "score" column.}
\usage{
setP(matches, motifs, bg)
}
\arguments{
\item{matches}{A GRangesList of matches from \code{\link{motifmatchr::matchMotifs}(out='positions')}}

\item{motifs}{The input motifs for matches.}

\item{bg}{Nucleotide background frequencies}
}
\value{
The GRangesList with p-values for a match in the metadata.
}
\description{
Add p-value to GRangesList with a "score" column.
}
\seealso{
\code{\link{TFMPvalue::TFMs2pv}}
}
